requirements.txt
shortuuid
django-jazzmin